Modularize javascript
