+++ src/js/modules/convertToMph.js
@@ -1,0 +1,5 @@
+import roundToTwo from './roundToTwoDecPlaces';
+
+const convertToMph = (kmh) => roundToTwo(kmh / 1.609);
+
+export default convertToMph;
