+++ src/js/modules/emptySearchBox.js
@@ -1,0 +1,5 @@
+const emptySearchBox = (inp) => {
+  inp.value = '';
+};
+
+export default emptySearchBox;
