+++ src/js/modules/errorPromptMessage.js
@@ -1,0 +1,7 @@
+const emptyInputMessage = document.querySelector('#empty-input-message');
+
+const errorPromptMessage = message => {
+  emptyInputMessage.textContent = message;
+};
+
+export default errorPromptMessage;
