+++ src/js/modules/fetchUserLocation.js
@@ -1,0 +1,8 @@
+import successCallBack from './successCallBack';
+import errorCallBack from './errorCallBack';
+
+const fetchUserLocation = () => {
+  navigator.geolocation.getCurrentPosition(successCallBack, errorCallBack);
+};
+
+export default fetchUserLocation;
