+++ src/js/modules/roundToTwoDecPlaces.js
@@ -1,0 +1,3 @@
+const roundToTwo = (num) => +(`${Math.round(`${num}e+2`)}e-2`);
+
+export default roundToTwo;
